Updated CPU temp function to return a double for precision

+++ cpu/cpu_temp.h
@@ -20,6 +20,6 @@
 #ifndef CPU_TEMP_H
 #define CPU_TEMP_H
 
-int get_cpu_temp(void);
+double get_cpu_temp(void);
 
 #endif
